+++ lib/mixin/mixin_with_database.dart
@@ -1,0 +1,13 @@
+import 'dart:convert';
+
+import 'package:wire/abstract/abstract_wire_database_service.dart';
+
+mixin WireMixinWithDatabase {
+  late final WireDatabaseServiceAbstract databaseService;
+
+  bool exist(String key)  => databaseService.exist(key);
+  Future<dynamic> retrieve(String key)  => databaseService.retrieve(key);
+  // ignore: empty_catches
+  void persist(String key, dynamic value) { try { databaseService.save(key, jsonEncode(value)); } catch (e) { } }
+  void delete(String key) => { if (databaseService.exist(key)) databaseService.delete(key) };
+}