+++ lib/mixin/mixin_with_when_ready.dart
@@ -1,0 +1,3 @@
+mixin WireMixinWithWhenReady {
+  late Future<bool> whenReady;
+}